requirements.txt
altair==5.5.0
appdirs==1.4.4
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
fastparquet==2024.11.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.10.0
nfl_data_py==0.3.3
numba==0.62.1
numpy==1.26.4
packaging==25.0
pandas==1.5.3
pillow==12.0.0
protobuf==6.33.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xgboost==3.1.1
